+++ src/_template.command.js
@@ -5,7 +5,7 @@ module.exports = class TemplateCommand extends SebediusCommand {
   constructor(client) {
     super(client, {
       ownerOnly: true, // <- Command is hidden & available for the bot owner only
-      category: SebediusCommand.CategoryFlagsBits.ROLL,
+      category: SebediusCommand.CategoryFlagsBits.UTILS,
       data: new SlashCommandBuilder()
         .setName('cmd')
         .setDescription('Template command'),
@@ -13,6 +13,7 @@ module.exports = class TemplateCommand extends SebediusCommand {
     });
   }
   /** @type {SebediusCommand.SebediusCommandRunFunction} */
+  // eslint-disable-next-line no-unused-vars
   async run(interaction, t, guildOptions) {
     // Add the command effects here.
   }