+++ src/commands/roll/roll.js
@@ -22,16 +22,18 @@ const Logger = require('../../utils/logger');
 const GameSubcommandsList = {
   [YearZeroGames.BLANK]: ['input', 'title', 'private'],
   [YearZeroGames.ALIEN_RPG]: [
-    'dice', 'stress', 'title', 'modifier', 'maxpush', 'private', 'fullauto', 'nerves', 'minpanic',
+    'dice', 'stress', 'title', 'modifier', 'maxpush', 'fullauto', 'private', 'nerves', 'minpanic',
   ],
   [YearZeroGames.BLADE_RUNNER]: ['abcd', 'title', 'modifier', 'maxpush', 'private'],
-  [YearZeroGames.CORIOLIS]: ['dice', 'title', 'modifier', 'maxpush', 'private', 'fullauto'],
+  [YearZeroGames.CORIOLIS]: ['dice', 'title', 'modifier', 'maxpush', 'fullauto', 'private'],
   [YearZeroGames.FORBIDDEN_LANDS]: [
-    'base', 'skill', 'gear', 'artifacts', 'title', 'modifier', 'maxpush', 'private', 'fullauto', 'pride',
+    'base', 'skill', 'gear', 'artifacts', 'title', 'modifier', 'maxpush', 'private', 'pride',
   ],
-  [YearZeroGames.MUTANT_YEAR_ZERO]: ['base', 'skill', 'gear', 'artifacts', 'title', 'modifier', 'maxpush', 'private'],
+  [YearZeroGames.MUTANT_YEAR_ZERO]: [
+    'base', 'skill', 'gear', 'artifacts', 'title', 'modifier', 'maxpush', 'fullauto', 'private',
+  ],
   [YearZeroGames.TALES_FROM_THE_LOOP]: ['dice', 'title', 'modifier', 'maxpush', 'private'],
-  [YearZeroGames.TWILIGHT_2K]: ['abcd', 'ammo', 'title', 'modifier', 'maxpush', 'private', 'fullauto'],
+  [YearZeroGames.TWILIGHT_2K]: ['abcd', 'ammo', 'title', 'modifier', 'maxpush', 'fullauto', 'private'],
   [YearZeroGames.VAESEN]: ['dice', 'title', 'modifier', 'maxpush', 'private'],
 };
 
@@ -43,7 +45,7 @@ const SlashCommandOptions = {
     required: true,
   },
   abcd: {
-    description: 'Type any of the following: 12, 10, 8, 6, a, b, c, d',
+    description: 'Type any of the following: `12`, `10`, `8`, `6`, `a`, `b`, `c`, `d`',
     type: ApplicationCommandOptionType.String,
     required: true,
   },
@@ -115,7 +117,7 @@ const SlashCommandOptions = {
     type: ApplicationCommandOptionType.Boolean,
   },
   minpanic: {
-    description: 'Adjusts a minimum treshold for multiple consecutive panic effects',
+    description: 'Adjust a minimum treshold for multiple consecutive panic effects',
     type: ApplicationCommandOptionType.Integer,
   },
 };
@@ -133,7 +135,8 @@ const SlashCommandOptions = {
 function _getSlashCommandBuilder() {
   const data = new SlashCommandBuilder()
     .setName('roll')
-    .setDescription('Roll dice for any Year Zero roleplaying game');
+    .setDescription('Roll dice for any Year Zero roleplaying game')
+    .setDMPermission(false);
 
   for (const [game, options] of Object.entries(GameSubcommandsList)) {
     data.addSubcommand(sub => {
@@ -273,7 +276,8 @@ module.exports = class RollCommand extends SebediusCommand {
 
     // Sets the maximum number of pushes.
     // Note: Cannot use typeof undefined because null is returned by DiscordJS
-    if (maxPush != null) roll.setMaxPush(maxPush);
+    if (this.isPanic(roll)) roll.setMaxPush(0);
+    else if (maxPush != null) roll.setMaxPush(maxPush);
     else {
       const fullauto = interaction.options.getBoolean('fullauto');
       if (fullauto) roll.setMaxPush(10);
@@ -281,13 +285,16 @@ module.exports = class RollCommand extends SebediusCommand {
 
     // Renders the roll to the chat.
     /** @type {import('discord.js').InteractionReplyOptions} */
-    const messageData = await this.render(roll, interaction, t);
+    const messageData = await this.render(roll, t);
     await interaction.reply({
       ...messageData,
       ephemeral: interaction.options.getBoolean('private'),
       fetchReply: true,
     });
 
+
+    // Detects panic.
+    if (this.isPanic(roll)) return this.fetchPanic(roll, interaction, t);
     if (roll.pushable) await this.awaitPush(roll, interaction, t);
   }
 
@@ -299,7 +306,7 @@ module.exports = class RollCommand extends SebediusCommand {
    * Parses & renders a generic roll.
    * @param {string} input
    * @param {string} title
-   * @param {import('discord.js').ChatInputCommandInteraction} interaction
+   * @param {SebediusCommand.SebediusCommandInteraction} interaction
    * @param {SebediusCommand.SebediusTranslationCallback} t
    */
   async executeGenericRoll(input, title, interaction, t) {
@@ -346,10 +353,9 @@ module.exports = class RollCommand extends SebediusCommand {
   /**
    * Renders the results of the roll into a chat message.
    * @param {YearZeroRoll} roll
-   * @param {import('discord.js').ChatInputCommandInteraction} interaction
    * @param {SebediusCommand.SebediusTranslationCallback} t
    */
-  async render(roll, interaction, t) {
+  async render(roll, t) {
     /** @type {import('$config').DiceRenderOptions} */
     const options = this.bot.config.Commands.roll.options[roll.game];
     if (!options) throw new ReferenceError(`[roll:${roll.game}] Command Options Not Found!`);
@@ -401,11 +407,12 @@ module.exports = class RollCommand extends SebediusCommand {
 
   /**
    * @param {YearZeroRoll} roll
-   * @param {import('discord.js').ChatInputCommandInteraction} interaction
+   * @param {SebediusCommand.SebediusCommandInteraction} interaction
    * @param {SebediusCommand.SebediusTranslationCallback} t
    */
   async awaitPush(roll, interaction, t) {
     const message = await interaction.fetchReply();
+    const gameOptions = this.bot.config.Commands.roll.options[roll.game];
 
     const collector = message.createMessageComponentCollector({
       componentType: ComponentType.Button,
@@ -420,14 +427,23 @@ module.exports = class RollCommand extends SebediusCommand {
           ephemeral: true,
         });
       }
-      else if (i.customId === 'push-button') {
+      else if (['push-button', 'pray-button', 'chapel-button'].includes(i.customId)) {
         await roll.push(true);
 
-        // Add any extra pushed dice
-        if (this.bot.config.Commands.roll.options[roll.game]?.extraPushDice) {
-          for (const ExtraDie of this.bot.config.Commands.roll.options[roll.game].extraPushDice) {
-            roll.addDice(ExtraDie, 1);
-          }
+        const extraPushDice = [];
+
+        // Adds any extra pushed dice.
+        if (gameOptions?.extraPushDice) {
+          extraPushDice.push(...gameOptions.extraPushDice);
+        }
+
+        // Idem, but for Coriolis.
+        if (['pray-button', 'chapel-button'].includes(i.customId)) {
+          extraPushDice.push(...gameOptions.pushMenu.find(b => b.customId === i.customId).extraPushDice);
+        }
+
+        if (extraPushDice.length) {
+          for (const ExtraDie of extraPushDice) roll.addDice(ExtraDie, 1);
           await roll.roll();
         }
 
@@ -445,7 +461,7 @@ module.exports = class RollCommand extends SebediusCommand {
         if (!roll.pushable) collector.stop();
 
         // Edits the message with the pushed roll.
-        const messageData = await this.render(roll, interaction, t);
+        const messageData = await this.render(roll, t);
 
         // TODO https://github.com/discordjs/discord.js/issues/8588
         // TODO https://github.com/discord/discord-api-docs/issues/5279
@@ -460,11 +476,9 @@ module.exports = class RollCommand extends SebediusCommand {
         }
 
         // Detects panic.
-        if (this.bot.config.Commands.roll.options[roll.game]?.panic && roll.panic) {
+        if (this.isPanic(roll)) {
           collector.stop();
-          const panicCommand = this.bot.commands.get('panic');
-          // TODO call PanicCommand
-          // await panicCommand.run(interaction, t);
+          await this.fetchPanic(roll, interaction, t);
         }
       }
       else if (i.customId === 'cancel-button') {
@@ -484,26 +498,73 @@ module.exports = class RollCommand extends SebediusCommand {
   /* ------------------------------------------ */
 
   #createButtons(game, t) {
+    const gameOptions = this.bot.config.Commands.roll.options[game];
+
     const pushButton = new ButtonBuilder()
       .setCustomId('push-button')
-      .setEmoji(this.bot.config.Commands.roll.options[game]?.successIcon || this.bot.config.Commands.roll.pushIcon)
-      .setLabel(t('commands:roll.button.push'))
+      .setEmoji(gameOptions?.successIcon || this.bot.config.Commands.roll.pushIcon)
+      .setLabel(t('commands:roll.buttons.push'))
       .setStyle(ButtonStyle.Primary);
 
     const cancelButton = new ButtonBuilder()
       .setCustomId('cancel-button')
       .setEmoji(this.bot.config.Commands.roll.cancelIcon)
-      .setLabel(t('commands:roll.button.cancel'))
+      .setLabel(t('commands:roll.buttons.cancel'))
       .setStyle(ButtonStyle.Secondary);
+
+    const actionRows = [];
 
     const firstActionRow = new ActionRowBuilder()
       .addComponents(pushButton, cancelButton);
 
-    return [firstActionRow];
+    actionRows.push(firstActionRow);
+
+    // If we have a custom push menu, we add the buttons configured in the config.
+    if (gameOptions?.pushMenu) {
+      const secondActionRow = new ActionRowBuilder();
+      for (const buttonData of gameOptions.pushMenu) {
+        const button = new ButtonBuilder({
+          style: ButtonStyle.Primary,
+          customId: buttonData.customId,
+          emoji: buttonData.emoji,
+          label: t(buttonData.label),
+        });
+        secondActionRow.addComponents(button);
+      }
+      actionRows.push(secondActionRow);
+    }
+
+    return actionRows;
   }
 
   /* ------------------------------------------ */
   /*  Utils Methods                             */
+  /* ------------------------------------------ */
+
+  /**
+   * @param {YearZeroRoll} roll
+   * @param {SebediusCommand.SebediusCommandInteraction} interaction
+   * @param {SebediusCommand.SebediusTranslationCallback} t
+   */
+  async fetchPanic(roll, interaction, t) {
+    Logger.roll('roll:alien → Panic!');
+    /** @type {import('./panic')} */
+    // @ts-ignore
+    const panicCommand = this.bot.commands.get('panic');
+    return panicCommand.panic(interaction, t, {
+      stress: roll.stress,
+      minPanic: interaction.options.getInteger('minpanic'),
+      isFixed: false,
+      hasNerves: interaction.options.getBoolean('nerves'),
+    });
+  }
+
+  /* ------------------------------------------ */
+
+  isPanic(roll) {
+    return this.bot.config.Commands.roll.options[roll.game]?.panic && roll.panic;
+  }
+
   /* ------------------------------------------ */
 
   /**