+++ src/structures/database/database.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose');
 const Schemas = require('./models');
 const Logger = require('../../utils/logger');
-const { isObjectEmpty } = require('../../utils/object-utils');
+// const { isObjectEmpty } = require('../../utils/object-utils'); // TODO clean
 
 class Database {
   constructor(client, uri) {
@@ -44,42 +44,71 @@ class Database {
       lean: true,
     });
   }
+  // TODO clean code
+  // /**
+  //  * See {@link Database.grab}
+  //  * @param {string}      id
+  //  * @param {UpdateData} [updateData]
+  //  */
+  // async grabGuild(id, updateData) {
+  //   return this.grab('guilds', id, updateData);
+  // }
 
-  /**
-   * @param {string|import('discord.js').Guild} g
-   * @param {UpdateData} [updateData]
-   */
-  async grabGuild(g, updateData) {
-    return this.grab('guilds', g, updateData);
-  }
+  // async getGuild(id) {
+  //   const document = await this.guilds.findById(id, null, { upsert: true });
+  //   if (document.isNew) Logger.client(`✨ Database | create: Guild ${id}`);
+  //   return document;
+  // }
 
-  /**
-   * @param {string} collection
-   * @param {string|import('discord.js').Guild} item
-   * @param {UpdateData} [updateData]
-   */
-  async grab(collection, item, updateData = {}) {
-    /** @type {typeof Schemas.Guild} */
-    const model = this[collection];
-    const id = typeof item === 'string' ? item : item.id;
-    let document = await model.findOne({ _id: id });
-    if (!document) {
-      document = new model({ ...updateData, _id: id });
-      await document.save();
-      Logger.client(`✨ Database | create: Guild ${id}`);
-      return document;
-    }
-    if (isObjectEmpty(updateData)) return document;
-    const data = {};
-    for (const [k, v] of Object.entries(updateData)) {
-      if (k !== 'id' && document[k] !== v) data[k] = v;
-    }
-    if (!isObjectEmpty(data)) {
-      await document.updateOne(data);
-      Logger.client(`📝 Database | update: Guild ${id} with ${JSON.stringify(data)}`);
-    }
-    return document;
-  }
+  // /**
+  //  * Returns a document from a collection, with the following:
+  //  * - Updates the document if update data is provided.
+  //  * - Creates the document (with the update data) if it does not exist.
+  //  * @param {string}      collection
+  //  * @param {string}      id
+  //  * @param {UpdateData} [updateData]
+  //  */
+  // async grab(collection, id, updateData = {}) {
+  //   /** @type {typeof Schemas.Guild} */
+  //   const model = this[collection];
+  //   if (!model) throw new TypeError(`Collection "${collection} does not exist in database!"`);
+  //   if (!id || typeof id !== 'string') return undefined;
+
+  //   const document = await model.findByIdAndUpdate(id, updateData, {
+  //     upsert: true,
+  //     new: true,
+  //     // maxTimeMS: 2000,
+  //   });
+
+  //   if (document.isNew) Logger.client(`✨ Database | create: Guild ${id}`);
+  //   if (document.isModified()) Logger.client(`📝 Database | update: Guild ${id} with ${JSON.stringify(updateData)}`);
+
+  //   return document;
+
+  //   // let document = await model.findOne({ _id: id });
+
+  //   // if (!document) {
+  //   //   document = new model({ ...updateData, _id: id });
+  //   //   await document.save();
+  //   //   Logger.client(`✨ Database | create: Guild ${id}`);
+  //   //   return document;
+  //   // }
+
+  //   // if (isObjectEmpty(updateData)) return document;
+
+  //   // const data = {};
+
+  //   // for (const [k, v] of Object.entries(updateData)) {
+  //   //   if (k !== 'id' && document[k] !== v) data[k] = v;
+  //   // }
+
+  //   // if (!isObjectEmpty(data)) {
+  //   //   await document.updateOne(data);
+  //   //   Logger.client(`📝 Database | update: Guild ${id} with ${JSON.stringify(data)}`);
+  //   // }
+
+  //   // return document;
+  // }
 }
 
 module.exports = Database;