+++ src/structures/sebedius-client.js
@@ -5,7 +5,7 @@ const Logger = require('../utils/logger');
 const handleEvents = require('./handlers/event-handler');
 const handleCommands = require('./handlers/command-handler');
 const loadLocales = require('../locales/i18n');
-const { Emojis } = require('../config');
+const { parseGamedata } = require('../yearzero/gamedata/gamedata-parser');
 
 class Sebedius extends Client {
   constructor(options) {
@@ -25,6 +25,9 @@ class Sebedius extends Client {
 
     /** @type {NodeJS.Timeout} */
     this.activity = null;
+
+    /** @type {import('i18next').default} */
+    this.i18n = null;
   }
 
   /* ------------------------------------------ */
@@ -60,11 +63,23 @@ class Sebedius extends Client {
   /*  Load Sebedius                             */
   /* ------------------------------------------ */
 
+  /**
+   * Starts the Sebedius client:
+   * - Connect the database
+   * - Load i18next
+   * - Load events & commands
+   * - Login to Discord
+   * @param {Object} options Options for starting Sebedius
+   * @param {string} options.dbURI    Database URI
+   * @param {string} options.events   Glob pattern for event files
+   * @param {string} options.commands Glob pattern for command files
+   * @param {string} options.token    Discord token
+   */
   async startSebedius(options) {
     // Attaches a database.
     this.database = new Database(this, options.dbURI);
 
-    await loadLocales();
+    await loadLocales(this);
 
     // Grafts our events & commands handlers to the client.
     await handleEvents(this, options.events);
@@ -79,7 +94,8 @@ class Sebedius extends Client {
   /* ------------------------------------------ */
 
   async getGuildsCount() {
-    return (await this.guilds.fetch()).size;
+    await this.guilds.fetch();
+    return this.guilds.cache.size;
   }
 
   async getUsersCount() {
@@ -111,8 +127,22 @@ class Sebedius extends Client {
   }
 
   async leaveBanned(guild) {
-    Logger.warn(`${Emojis.ban} Guild is banned! → Leaving...`);
+    Logger.warn(`${this.config.Emojis.ban} Guild is banned! → Leaving...`);
     return guild.leave();
+  }
+
+  /* ------------------------------------------ */
+  /*  Table Methods                             */
+  /* ------------------------------------------ */
+
+  /**
+   * 
+   * @param {import('discord.js').LocaleString} locale 
+   * @param {string} name 
+   * @returns {Promise.<import('@utils/RollTable')>}
+   */
+  async getTable(locale, name) {
+    return parseGamedata(locale, name);
   }
 }